requirements.txt
discord.py
aiohttp
tzdata
requests